Extract Common Portions of AddJob & EditJob
